# resources/swagger/oauth2-swagger.yml
swagger: "2.0"
info:
  version: "1.0"
  title: "Kong Adapter OAuth 2.0 Helper API"
basePath: "/"
consumes:
  - application/json
schemes:
  - http
paths:
  /oauth2/token/implicit:
    post:
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/implicitTokenType'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/implicitSuccessType'
        401:
          description: Unauthorized
        404:
          description: Not found

  /oauth2/token/password:
    post:
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/passwordTokenType'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/accessTokenType'
        401:
          description: Unauthorized
        404:
          description: Not found

  /oauth2/token/refresh:
    post:
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/refreshTokenType'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/accessTokenType'
        400:
          description: Bad request
        404:
          description: Not found

  /oauth2/token:
    get:
      parameters:
        - name: access_token
          in: query
          type: string
          description: Pass in either `access_token` or `refresh_token` to retrieve token data
        - name: refresh_token
          in: query
          type: string
          description: Pass in either `access_token` or `refresh_token` to retrieve token data
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/accessTokenType'
        404:
          description: Not found
  delete:
    parameters:
      - name: authenticated_userid
        in: query
        type: string
        description: Pass in either `access_token` or `authenticated_userid` to delete token(s)
      - name: access_token
        in: query
        type: string
        description: Pass in either `access_token` or `authenticated_userid` to delete token(s)
    responses:
      204:
        description: Successfully delete, no response
      404:
        description: Not found
definitions:
  implicitTokenType:
    type: object
    properties:
      authenticated_userid:
        type: string
      api_id:
        type: string
      client_id:
        type: string
      scope:
        type: array
        items:
          type: string
  implicitSuccessType:
    type: object
    properties:
      redirect_uri:
        type: string

  passwordTokenType:
    type: object
    properties:
      authenticated_userid:
        type: string
      api_id:
        type: string
      client_id:
        type: string
      scope:
        type: array
        items:
          type: string
  
  accessTokenType:
    type: object
    properties:
      access_token:
        type: string
      token_type:
        type: string
      refresh_token:
        type: string
      expires_in:
        type: number

  refreshTokenType:
    type: object
    properties:
      refresh_token:
        type: string
      client_id:
        type: string